modemmanager: Add 1.6.8 recipe for jethro builds
Currently oe provides an older version in fido. This recipe is based on the
one from meta-oe (morty).

Inheriting gobject-introspection was removed as it is not available in jethro.

Jethro doesn't have the bash-completion class and not using it breaks the build
as there are files unpackaged. Fix this by adding in recipe support for
packaging these files as the new class does.

Signed-off-by: Andrei Gherzan <andrei@resin.io>