libmbim: Update to 1.14.0 in jethro
Updating modem manager needs an update to its libmbim dependency too. Make sure
this satisfy the following configure requirement:

    configure: error: Couldn't find libmbim-glib >= 1.14.0. Install it, or
    otherwise configure using --without-mbim to disable MBIM support.

Jethro doesn't have the bash-completion class and not using it breaks the build
as there are files unpackaged. Fix this by adding in recipe support for
packaging these files as the new class does.

Signed-off-by: Andrei Gherzan <andrei@resin.io>