libqmi: Update to 1.16.0 in jethro
The update of modem manager to 1.6.8 requires an updated libqmi too:

    configure: error: Couldn't find libqmi-glib >= 1.16.0. Install it, or otherwise
    configure using --without-qmi to disable QMI support.

The recipe is based on the one from meta-oe (morty).

Jethro doesn't have the bash-completion class and not using it breaks the build
as there are files unpackaged. Fix this by adding in recipe support for
packaging these files as the new class does.

Signed-off-by: Andrei Gherzan <andrei@resin.io>